+++ core/src/main/java/io/substrait/expression/AbstractExpressionVisitor.java
@@ -202,6 +202,11 @@ public abstract class AbstractExpressionVisitor<O, C extends VisitationContext, 
   }
 
   @Override
+  public O visit(Expression.NestedList expr, C context) throws E {
+    return visitFallback(expr, context);
+  }
+
+  @Override
   public O visit(FieldReference expr, C context) throws E {
     return visitFallback(expr, context);
   }